work on har

+++ classification/har/adbox.py
@@ -13,9 +13,9 @@ class ADBox(td.Dataset):
 
     def __getitem__(self, index: int):
         file = self.files[index]
-        depth = np.load(self.data_root + '/depth_feature/' + file + '.npy')
-        radar = np.load(self.data_root + '/radar/' + file + '.npy')
-        imu = pd.read_csv(self.data_root + '/imu/' + file + '.csv').to_numpy()[:, :-1]
+        depth = np.load(self.data_root + '/depth_feature/' + file + '.npy').astype(np.float32)
+        radar = np.load(self.data_root + '/radar/' + file + '.npy').astype(np.float32)
+        imu = pd.read_csv(self.data_root + '/imu/' + file + '.csv').to_numpy().astype(np.float32)[:, :-1]
         label = self.label[index]
         print(depth.shape, radar.shape, imu.shape, label)
         print(type(depth), type(radar), type(imu), type(label))
